Work on status
Set of updates.
The main things are :
- The "callé" attrbute is no more used
- "The status has 3 states : 0 -> not ready, 1 -> ready; 2 -> finished.
This applies to drives as welll as to runs. He runs status is automatically calculated according to the statuses of its drives (the min status is kept). As a consequence, the status of the run can not be defined in the run's pages (list, modif, creation).
Other modifications :
- the list of runs for a group is available on the groups page (display and modif)
- a drive can be removed directly from the timeline as soon as at least one remains for the run. If not, the run itself must be removed (accessible from the runs'list). The reason is a drive can to esaily be removed clicking on the red cross.